Moved back to maintain docs and test packages in requirements files
Maintaining the packages to build the docs and run the tests in requirement
files is more flexible than using setup.py. It is not possible to add forks to
setup.py requirements, but requirement files support this.

requirements/forks.txt has been introduced to allow forks of packages required in setup.py.

In addition dependencies in tox.ini have been pinned.